Moves watchers to session
[ui tests]
